feat: DB data type 변경에 따른 entity data type 변경, 사일로 작업을 위한 db 구축

+++ src/main/java/org/cric/back_office/global/config/JwtAuthenticationFilter.java
@@ -36,7 +36,7 @@ public class JwtAuthenticationFilter extends OncePerRequestFilter {
         System.out.println(token);
         if (token != null) {
             if (jwtUtil.validateToken(token)) {
-                Integer userId = jwtUtil.getUserIdFromToken(token);
+                Long userId = jwtUtil.getUserIdFromToken(token);
                 String tokenId = jwtUtil.getTokenIdFromToken(token);
 
                 // Redis에서 토큰 ID 유효성 검증 (중복 로그인 방지)
